requirements.txt
semantic-kernel
openai
python-dotenv
fastapi
uvicorn
python-multipart
